server/pyproject.toml
[tool.poetry]
name = "app-warframe-toolkit"
version = "0.1.0"
description = ""
authors = ["Eric Batista <dev.eric.batista@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
eric-devtools = "^0.3.2"
uvicorn = "^0.29.0"
alembic = "^1.13.1"
psycopg2 = "^2.9.9"
fastapi = "^0.111.0"


[tool.poetry.group.style.dependencies]
black = "^24.4.2"
isort = "^5.13.2"
autoflake = "^2.3.1"


[tool.poetry.group.test.dependencies]
pytest = "^8.2.0"
pytest-cov = "^5.0.0"

#[tool.pytest.ini_options]
#addopts = [
#    "--cov=warframe_manager/",
#    "--cov-report=json",
#    "--cov-config=.coveragerc",
#    "--no-cov-on-fail",
#    "--cov-fail-under=80",
#    "--nf",
#    "-s",
#    "--maxfail=1",
#]
#filterwarnings = [
#    "error",
#    "ignore:The --rsyncdir command line argument and rsyncdirs config variable are deprecated.:DeprecationWarning",
#]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
